Hash utility functions were extracted to a separate module for reuse
